Clean things up from hack session
